requirements.txt
discord.py
python-dotenv
requests
jishaku
discord
datetime